Split lib.rs into serialization and deserialization
